ipmi: call check_timers() while waiting for synchronous messages to complete
BT responses are handled using a timer doing the polling. To hope to
get an answer to an IPMI synchronous message, the timer needs to run.

This issue shows up very quickly under QEMU when loading the first
flash resource with the IPMI HIOMAP backend.

Adding a timeout would also help in reporting errors instead of
looping indefinitely waiting for a response.

Signed-off-by: Cédric Le Goater <clg@kaod.org>