libstb/tss: implement PCR Extend command for TPM 1.2 as well
Only SHA-1 is supported by TPM 1.2, so the function will pick that hash
from those passed to tpmCmdPcrExtend2Hash() and fail if SHA-1 is not
present.

This is the only command actually used by skiboot.  The only other one
that can be issued is PCR Read, but that's done only for debug dumps.
Several others which are implemented are never invoked.

Signed-off-by: Sergii Dmytruk <sergii.dmytruk@3mdeb.com>